Allowed decimal input for temperature (because maybe the user made a bet with friends that 14.9 degrees celsius is cold.

+++ js/script.js
@@ -10,7 +10,7 @@ function judge() {
   const SUPERNOVA = 1000000000000
   
   // Variables
-  let celsius = parseInt(document.getElementById('celsius').value) // Makes variable an integer
+  let celsius = parseFloat(document.getElementById('celsius').value) // Makes variable an integer
 
   // Output messages if the user got it or not
   let message = ""
